styling different errors

+++ app.py
@@ -53,7 +53,7 @@ def add_department():
         data = (dept_name, dept_total)
         db_connection = db.connect_to_database()
         execute_query(db_connection, query, data)
-        flash('department add was successful', 'info')
+        flash('department add was successful', 'success')
         # below we display a success page (instead we need to redirect to 'departments' page and add flash message here.
         # return render_template('success.html', action= 'Add department', entity='Departments', active='departments', return_page='/departments')  retuns a success message
         return redirect('/departments')
@@ -77,7 +77,7 @@ def update_department():
         data = (dept_name, dept_id)
         execute_query(db_connection, query, data)
         # add flash message here if update was successful
-        flash('update department was successful')
+        flash('update department was successful', 'successful')
         return redirect('/departments')     # after update redirect to the page to show results
 
 
@@ -86,12 +86,18 @@ def delete_department():
     """ Deletes a department from the database. This function expects an integer argument named 'dept_id' """
     dept_id = request.args.get('dept_id')
     db_connection = db.connect_to_database()
+    query = "SELECT dept_total FROM Departments WHERE dept_id = %s;" % (dept_id)
+    dept = (db_connection, query).fetchone()
+    if dept['dept_total'] > 0:
+        flash("can't delete a department with employees in it, please reassign employees", "error")
+        return redirect('/departments')
+
     query = "DELETE from Departments WHERE dept_id = %s ;"
     data = (dept_id,)
     execute_query(db_connection, query, data)
     # add flash message here if delete was successful
-    flash('department delete was successful')
-    return redirect('/departments')        # after delete redirect to page to show results
+    flash('department delete was successful', 'successful')
+    return redirect('/departments')  # after delete redirect to page to show results
 
 
 # -----------End Departments------------------------------------------------------------
@@ -123,7 +129,7 @@ def employee_search():
     if results:
         return render_template('employee_search.html', entity=results)
     else:
-        flash('No results found for entry')
+        flash('No results found for entry', 'warning')
         return redirect('/employees')
 
 
@@ -169,7 +175,7 @@ def add_employee():
         dept_number)
         execute_query(db_connection, query, data)
         # need to return to redirected 'employees' page and add flash message here
-        flash('employee added successfully')
+        flash('employee added successfully', 'successful')
         # return render_template('success.html', action = 'Add employee', entity='Employees', active='employees', return_page='employees') # returns a success message
         return redirect('/employees')
 
@@ -213,7 +219,7 @@ def update_employee():
         data = (f_name, l_name, gender, address_1, address_2, city, state, zip, tel, email, start_date, end_date, status, dept_number, emp_id)
         execute_query(db_connection, query, data)
         # add flash message here
-        flash('employee update successful')
+        flash('employee update successful', 'successful')
         return redirect('/employees')
 
 
@@ -226,7 +232,7 @@ def delete_employee():
     data = (emp_id,)
     execute_query(db_connection, query, data)
     # flash message here
-    flash('employee delete successful')
+    flash('employee delete successful', 'successful')
     return redirect('/employees')
 
 # -------------End Employees----------------------------------------------------------
@@ -257,7 +263,7 @@ def add_certification():
         data = (cert_name)
         db_connection = db.connect_to_database()
         execute_query(db_connection, query, data)
-        flash('certification add was successful')
+        flash('certification add was successful', 'successful')
         # return render_template('success.html', action = 'Add certification', entity='Certifications', active='certifications', return_page='certifications') # retuns a success message
         return redirect('/certifications')
 
@@ -279,7 +285,7 @@ def update_certification():
         query = "UPDATE Certifications SET cert_name = %s WHERE cert_id = %s ;"
         data = (cert_name, cert_id)
         execute_query(db_connection, query, data)
-        flash('update certification was successful')
+        flash('update certification was successful', 'successful')
         return redirect('/certifications')
 
 
@@ -290,7 +296,7 @@ def delete_certification():
     query = "DELETE from Certifications WHERE cert_id = %s ;"
     data = (cert_id,)
     execute_query(db_connection, query, data)
-    flash('certification was successfully deleted')
+    flash('certification was successfully deleted', 'successful')
     return redirect('/certifications')
 
 # -----------End Certifications------------------------------------------------------------
@@ -332,7 +338,7 @@ def add_job():
         # put data in a tuple and execute query
         data = (dept_number, job_description)
         execute_query(db_connection, query, data)
-        flash('job add was successful')
+        flash('job add was successful', 'successful')
         # return render_template('success.html', action = 'Add job', entity='Jobs', active='jobs', return_page='jobs')
         # returns a success message
         return redirect('/jobs')
@@ -358,7 +364,7 @@ def update_job():
         query = "UPDATE Jobs SET job_description = %s, dept_number = %s WHERE job_id = %s ;"
         data = (job_description, dept_id, job_id)
         execute_query(db_connection, query, data)
-        flash('job updated successfully')
+        flash('job updated successfully', 'successful')
         return redirect('/jobs')
 
 
@@ -369,7 +375,7 @@ def delete_job():
     query = "DELETE from Jobs WHERE job_id = %s ;"
     data = (job_id,)
     execute_query(db_connection, query, data)
-    flash('job was deleted successfully')
+    flash('job was deleted successfully', 'successful')
     return redirect('/jobs')
 
 # ---------------End Jobs--------------------------------------------------------
@@ -396,7 +402,7 @@ def add_class():
     if request.method == 'GET':
         query = "SELECT emp_id, f_name, l_name FROM Employees WHERE dept_number = 1 or dept_number = 2;"
         results = db.execute_query(db_connection=db_connection, query=query).fetchall()
-        flash('class add was successful')
+        flash('class add was successful', 'successful')
         return render_template('add_class.html', instructor=results)
 
     # runs if it needs to update the database
@@ -414,7 +420,7 @@ def add_class():
         # put data in a tuple and execute query
         data = (class_name, instructor, time, length, class_total, class_max)
         execute_query(db_connection, query, data)
-        flash('class add was successful')
+        flash('class add was successful', 'successful')
         # return render_template('success.html', action = 'Add class', entity='Classes', active='classes', return_page='classes') # retuns a success message
         return redirect('/classes')
 
@@ -445,7 +451,7 @@ def update_class():
         query = "UPDATE Classes SET class_name = %s, instructor = %s, time = %s, length = %s, class_total = %s, class_max = %s WHERE class_id = %s ;"
         data = (class_name, instructor, time, length, class_total, class_max, class_id)
         execute_query(db_connection, query, data)
-        flash('class update was successful')
+        flash('class update was successful', 'successful')
         return redirect('/classes')
 
 
@@ -456,7 +462,7 @@ def delete_class():
     query = "DELETE from Classes WHERE class_id = %s ;"
     data = (class_id,)
     execute_query(db_connection, query, data)
-    flash('class was successfully deleted')
+    flash('class was successfully deleted', 'successful')
     return redirect('/classes')
 # ---------------End Classes--------------------------------------------------------
 
@@ -487,7 +493,7 @@ def members_search():
     if results:
         return render_template('members_search.html', entity=results)
     else:
-        flash('No results found for entry')
+        flash('No results found for entry', 'warning')
         return redirect('/members')
 
 
@@ -520,7 +526,7 @@ def add_member():
     # connect to database
     db_connection = db.connect_to_database()
     execute_query(db_connection, query, data)
-    flash('member add was successful')
+    flash('member add was successful', 'successful')
     return redirect('/members')
     # return render_template('success.html', action ='Add member', entity='Members', active='members', return_page='members')
 
@@ -552,7 +558,7 @@ def update_member():
         query = "UPDATE Members SET f_name = %s, l_name = %s, gender = %s, address_1 = %s, address_2 = %s, city = %s, state = %s, zip = %s, tel = %s, email = %s, status = %s WHERE member_id = %s;"
         data = (f_name, l_name, gender, address_1, address_2, city, state, zip, tel, email, status, member_id)
         execute_query(db_connection, query, data)
-        flash('member update was successful')
+        flash('member update was successful', 'successful')
         return redirect('/members')
 
 
@@ -563,7 +569,7 @@ def delete_member():
     query = "DELETE from Members WHERE member_id = %s ;"
     data = (member_id,)
     execute_query(db_connection, query, data)
-    flash('member delete was successful')
+    flash('member delete was successful', 'successful')
     return redirect('/members')
 
 # ----------------End Members-------------------------------------------------------
@@ -608,7 +614,7 @@ def add_emp_certs():
         query = "INSERT INTO Emp_Certs (emp_id, cert_id) VALUES (%s,%s);"
         data = (employee, cert)
         execute_query(db_connection, query, data)
-        flash('employee certification add was successful')
+        flash('employee certification add was successful', 'successful')
         return redirect('/emp_certs')
         # return render_template('success.html', action='Certification added to an employee', entity='Emp_Certs', active='other', return_page='emp_certs')
 
@@ -638,7 +644,7 @@ def update_emp_certs():
         query = "UPDATE Emp_Certs SET emp_id = %s, cert_id = %s WHERE emp_id = %s;"
         data = (emp_id, cert_id, emp_id)
         execute_query(db_connection, query, data)
-        flash('employee certification update was successful')
+        flash('employee certification update was successful', 'successful')
         return redirect('emp_certs')
 
 
@@ -654,7 +660,7 @@ def delete_emp_certs():
     query = "DELETE from Emp_Certs WHERE emp_id = %s and cert_id = %s ;"
     data = (emp_id, cert_id)
     execute_query(db_connection, query, data)
-    flash('employee certification was successfully deleted')
+    flash('employee certification was successfully deleted', 'successful')
     return redirect('/emp_certs')
 
 # ------------------End Emp_Certs-----------------------------------------------------
@@ -691,12 +697,13 @@ def add_mem_classes():
         check_cap = cursor.fetchone()
         if check_cap['class_total'] == check_cap['class_max']:  # make sure class isn't full
             # insert flash message here (class is at capacity)
+            flash('class is full', 'error')  # implement category for filters
             print('class is full')
             return redirect('/mem_classes')
         query = "INSERT INTO Mem_Classes (member_id, class_id) VALUES (%s,%s);"
         data = (member, class_id)
         execute_query(db_connection, query, data)
-        flash("member's class was successfully added ")
+        flash("member's class was successfully added ", 'successful')
         return redirect('/mem_classes')
         # return render_template('success.html', action='Member add to class', entity='Mem_Classes', active='other', return_page='mem_classes')
 
@@ -723,12 +730,13 @@ def enroll_into_class():
         check_cap = cursor.fetchone()
         if check_cap['class_total'] == check_cap['class_max']:  # make sure class isn't full
             # insert flash message here (class is at capacity)
+            flash('sorry, this class is full', 'warning')
             print('class is full')
             return redirect('/mem_classes')
         query = "INSERT INTO Mem_Classes (member_id, class_id) VALUES (%s,%s);"
         data = (member, class_id)
         execute_query(db_connection, query, data)
-        flash("member's class was successfully added ")
+        flash("member's class was successfully added ", 'successful')
         return redirect('/mem_classes')
 
 
@@ -753,7 +761,7 @@ def update_mem_class():
         query = "UPDATE Mem_Classes SET member_id = %s, class_id = %s WHERE member_id = %s and class_id = %s;"
         data = (member_id, class_id, member_id, curr_class_id)
         execute_query(db_connection, query, data)
-        flash("member's class was successfully updated ")
+        flash("member's class was successfully updated ", 'successful')
         return redirect('mem_classes')
 
 
@@ -765,7 +773,7 @@ def delete_mem_class():
     query = "DELETE from Mem_Classes WHERE member_id = %s and class_id = %s ;"
     data = (member_id, class_id)
     execute_query(db_connection, query, data)
-    flash("member's class was successfully deleted")
+    flash("member's class was successfully deleted", 'successful')
     return redirect('/mem_classes')
 
 # -----------------End Mem_Classes------------------------------------------------------
@@ -804,7 +812,7 @@ def add_emp_jobs():
         query = "INSERT INTO Emp_Jobs (emp_id, job_id) VALUES (%s,%s);"
         data = (employee, job)
         execute_query(db_connection, query, data)
-        flash('employee was successfully assigned to a job')
+        flash('employee was successfully assigned to a job', 'successful')
         return redirect('/emp_jobs')
         # return render_template('success.html', action='Assignment was', entity='Emp_Jobs', active='other', return_page='emp_jobs')
 
@@ -830,7 +838,7 @@ def update_emp_job():
         query = "UPDATE Emp_Jobs SET emp_id = %s, job_id = %s WHERE emp_id = %s;"
         data = (emp_id, job_id, emp_id)
         execute_query(db_connection, query, data)
-        flash("employee's job was successfully updated")
+        flash("employee's job was successfully updated", 'successful')
         return redirect('emp_jobs')
 
 
@@ -842,7 +850,7 @@ def delete_emp_job():
     query = "DELETE from Emp_Jobs WHERE emp_id = %s and job_id = %s ;"
     data = (emp_id, job_id)
     execute_query(db_connection, query, data)
-    flash("employee's job was successfully deleted")
+    flash("employee's job was successfully deleted", 'successful')
     return redirect('/emp_jobs')
 
 
@@ -865,7 +873,7 @@ def job_emps():
     result = cursor.fetchall()
     if result is None:
         print(result)  # add flash message here
-        flash("employee isn't assigned to any jobs")
+        flash("employee isn't assigned to any jobs", 'warning')
         return redirect('/jobs')
     else:
         return render_template('job_emps.html', entity=result, job_id=job_id, page='Jobs', return_page='/jobs')
@@ -921,7 +929,7 @@ def emp_dept():
     print(results)
     if results == ():
         print(results)  # add flash message here
-        flash("this department doesn't have any employees")
+        flash("this department doesn't have any employees", 'warning')
         return redirect('/departments')
     else:
         return render_template('emp_dept.html', entity=results, dept_name=dept_name)
